RLogin-2.29.9

+++ RLogin/sntrup761.cpp
@@ -1783,8 +1783,8 @@ static void Encode(unsigned char *out, const uint16_t *R, const uint16_t *M, lon
   }
   if (len > 1) {
 //    uint16_t R2[(len + 1) / 2], M2[(len + 1) / 2];
-	uint16_t *R2 = new uint16_t[(len + 1) / 2];
-	uint16_t *M2 = new uint16_t[(len + 1) / 2];
+	uint16_t *R2 = new uint16_t[(int)((len + 1) / 2)];
+	uint16_t *M2 = new uint16_t[(int)((len + 1) / 2)];
     long long i;
     for (i = 0; i < len - 1; i += 2) {
       uint32_t m0 = M[i];
@@ -1821,10 +1821,10 @@ static void Decode(uint16_t *out, const unsigned char *S, const uint16_t *M, lon
   if (len > 1) {
     //uint16_t R2[(len + 1) / 2], M2[(len + 1) / 2], bottomr[len / 2];
     //uint32_t bottomt[len / 2];
-	uint16_t *R2 = new uint16_t[(len + 1) / 2];
-	uint16_t *M2 = new uint16_t[(len + 1) / 2];
-	uint16_t *bottomr = new uint16_t[len / 2];
-	uint32_t *bottomt = new uint32_t[len / 2];
+	uint16_t *R2 = new uint16_t[(int)((len + 1) / 2)];
+	uint16_t *M2 = new uint16_t[(int)((len + 1) / 2)];
+	uint16_t *bottomr = new uint16_t[(int)(len / 2)];
+	uint32_t *bottomt = new uint32_t[(int)(len / 2)];
     long long i;
     for (i = 0; i < len - 1; i += 2) {
       uint32_t m = M[i] * (uint32_t)M[i + 1];
